requirements.txt
autopep8==1.3.2
certifi==2017.7.27
cffi==1.10.0
chardet==3.0.4
contextlib2==0.5.5
Django==1.11.4
django-appconf==1.0.2
django-cacheops==3.2.1
django-compressor==2.2
django-debug-toolbar==1.8
django-silk==1.0.0
fastpbkdf2==0.2
freezegun==0.3.9
funcy==1.8
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.0
num2words==0.5.5
psycopg2==2.7.3
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
raven==6.1.0
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.3
urllib3==1.22
uWSGI==2.0.15
